refactor(ios): split extension controller view state
Extract the iOS extension KeyboardViewModel and KeyboardContainerView out of KeyboardViewController so the controller file stays focused on UIKit and delegate integration.

Update ERICK-141 and ERICK-144 to record that the code split is complete on this branch, while macOS Xcode validation is still pending outside this Windows environment.
